% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_htr_grapher.R
\name{bh_htr_grapher}
\alias{bh_htr_grapher}
\title{Produire le graphique, pour une station, du débit en fonction du temps en représentant
    QMNA5, module et normale mansuelle}
\usage{
bh_htr_grapher(
  debit_sh,
  synthese_sh,
  donnees_sh,
  coul_em = "#54b5c5",
  coul_QMNA5 = "red",
  coul_module = "brown",
  coul_actuel = "darkgreen"
)
}
\arguments{
\item{debit_sh}{Dataframe contenant les données de débit temps réel pour la station,
produit par la fonction bh_htr().}

\item{synthese_sh}{Dataframe contenant les données de débit synthèse pour la station,
produit par la fonction bh_sy_ind_sta().}

\item{donnees_sh}{Dataframe contenant les catactéristiques de la station,
produit par la fonction bh_sta_data()}

\item{coul_em}{Caractère. Code de la couleur associée aux écoulements mensuels.}

\item{coul_QMNA5}{Caractère. Code de la couleur associée au QMNA5.}

\item{coul_module}{Caractère. Code de la couleur associée au module.}

\item{coul_actuel}{Caractère. Code de la couleur associée aux débits temps réel eu aux
données les plus récentes.}
}
\value{
Un graphique de classe ggplot2.
}
\description{
Produire le graphique, pour une station, du débit en fonction du temps en représentant
    QMNA5, module et normale mansuelle
}
\examples{
\dontrun{
mon_graphique <- bh_htr_grapher(debit_sh = debit_tr,
synthese_sh = synthese_ma_sh,
donnees_sh = sh_data)
}
}
