% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_sy_ind_sta.R
\name{bh_sy_ind_sta}
\alias{bh_sy_ind_sta}
\title{Extraire un type d'indicateur statistique de la liste des synthèses}
\usage{
bh_sy_ind_sta(syntheses, sta_id, indicateur = "em")
}
\arguments{
\item{syntheses}{Liste des synthèses par station.}

\item{sta_id}{Caractère. Code station.}

\item{indicateur}{Caractère. Type d'indicateur statistique à extraire.}
}
\value{
Un dataframe avec la statistique pour la station. Par défaut ce sont les
    écoulements mensuels.
}
\description{
Extraire un type d'indicateur statistique de la liste des synthèses
}
\examples{
\dontrun{
bh_sy_ind_sta(syntheses, sta_id = "J1813010", stat = "em")
}
}
