% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_sy_parser.R
\name{bh_sy_parser}
\alias{bh_sy_parser}
\title{Parser l'export csv des synthèses de la banque HYDRO}
\usage{
bh_sy_parser(fichier)
}
\arguments{
\item{fichier}{Texte. Chemin vers le fichier d'export au format csv.}
}
\value{
Une liste de listes structurée. Elle contient autant d'éléments qu'il y a de
    stations dans les données. Pour chaque station, la liste comprend 8 tableaux
    correspondant aux débits mensuels, crues, modules, étiages, etc.
}
\description{
L'export est un empilement non structuré de tableaux. La fonction permet d'interpréter
    les données par station et de constituer pour chacune un tableau par thème (crues,
    étiages, etc.).
}
\examples{
\dontrun{
syntheses <- bh_sy_parser(fichier = "raw_data/5590_4_synthese.csv")
}
}
